// src/Modules/TPM/Constants/Interfaces.ts
export enum DocumentInfo {
    OTHER_DOCUMENT_TYPE_CODE = "M99",
    RELEASE_ORDER_DOCUMENT_TYPE_CODE = "D04",
    RELEASE_ORDER_COM_DOCUMENT_TYPE_CODE = "D03",
    GOODS_DECLARATION_DOCUMENT_TYPE_CODE = "C51",
    IMPORT_PERMIT_DOCUMENT_TYPE_CODE = "D11",
    IMPORT_PERMIT_AMENDMENT_DOCUMENT_TYPE_CODE = "D56",
    MNC_IMPORT_PERMIT_AMENDMENT_DOCUMENT_TYPE_CODE = "A32",
    IMPORT_PERMIT_COM_DOCUMENT_TYPE_CODE = "D12",
    TRADE_TRAN_TYPE = 1,
    EXPORT_CERTIFICATE_DOCUMENT_TYPE_CODE = "D06",
    EXPORT_CERTIFICATE_COM_DOCUMENT_TYPE_CODE = "D15",
    AQD_EXPORT_CERTIFICATE_DOCUMENT_TYPE_CODE = "A05",
    AQD_EXPORT_CERTIFICATE_COM_DOCUMENT_TYPE_CODE = "A06",
    PHYSICAL_INSPECTION_FUMIGATION_REPORT = "M88",
    PEST_RISK_ANALYSIS_CERTIFICATE_TYPE_CODE = "D10",
    PEST_RISK_ANALYSIS_CERTIFICATE_DOCUMENT_TYPE_NAME = "Pest Risk Analysis Certificate/Any Other Official Document/Correspondence reference",
    PHYSICAL_INSPECTION_FUMIGATION_REPORT_DOCUMENT_TYPE_NAME = "Physical Inspection Treatment Report",
    FSCRD_SEED_ENLISTMENT_DOCUMENT_TYPE_CODE = "SE1",
    FSCRD_SEED_ENLISTMENT_COMMODITY_DOCUMENT_TYPE_CODE = "SE2",
    FSCRD_SEED_BUSINESS_NATIONAL_REGISTER = "F27",
    FSCRD_SEED_BUSINESS_NATIONAL_REGISTER_CERTIFICATE = "F28",
    MFD_PREMISE_REGISTRATION_DOCUMENT_TYPE_CODE = "A09",
    MFD_PREMISE_AMENDMENT_DOCUMENT_TYPE_CODE = "A17",
    MFD_PREMISE_RENEWAL_DOCUMENT_TYPE_CODE = "A18",
    DRAP_ENTITY_REGISTRATION_DOCUMENT_TYPE_CODE = "D90",
    MNC_PRODUCT_REGISTRATION_DOCUMENT_TYPE_CODE = "A30",
    MNC_PRODUCT_REGISTRATION_CERT_DOCUMENT_TYPE_CODE = "A31",
    MNC_PRODUCT_REGISTRATION_AMNEDMENT_DOCUMENT_TYPE_CODE = "A33",
    DPP_PREMISE_REGISTRATION_DOCUMENT_TYPE_CODE_18 = "F30",
    DPP_PREMISE_REGISTRATION_DOCUMENT_TYPE_CODE_19 = "F32",
    DPP_PREMISE_REGISTRATION_DOCUMENT_TYPE_CODE_20 = "F34",
    EXPORT_CERTIFICATE_EXTENSION = "E02",
    EXPORT_CERTIFICATE_MFD_EXTENSION = "E03",
    EXPORT_CERTIFICATE_REVIEW = "R50",
    SINGLE_DECLARATION_DOCUMENT_TYPE_CODE = "PS1",
    DRAP_DRUG_MANUFACTURING_LICENSE_REGISTRATION = "D92",
    DRAP_DRUG_REGISTRATION = "D93",
    DRAP_DRUG_IMPORT_LICENSE_REGISTRATION = "D91",
    IMPORT_PERMIT_DRAP_OPINION_REQUEST = "D86",
    IMPORT_PERMIT_FBR_OPINION_REQUEST = "D87",
    MNC_PRODUCT_GROUND_CHECK_REPORT_DOCUMENT_TYPE_CODE = "G01",

    //Treatment TPM
    TREATMENT_REG_DOC_TYPE_CODE = "C09",
    TREATMENT_REG_AMENDMENT_DOC_TYPE_CODE = "C11",
    TREATMENT_REG_RENEWAL_DOC_TYPE_CODE = "C13",

    // PSQCA
    PSQCA_RELEASE_ORDER_DOCUMENT_TYPE_CODE = "R01",
    PSQCA_RELEASE_ORDER_REQUEST_DOCUMENT_TYPE_CODE = "R02",

    FSCRD_RELEASE_ORDER_DOCUMENT_TYPE_CODE = "R05",
    FSCRD_RELEASE_ORDER_COM_DOCUMENT_TYPE_CODE = "R04",
    DPP_PREMISE_FORM_18_REGISTRATION_DOCUMENT_TYPE_CODE = "F30",
    DPP_PREMISE_FORM_19_REGISTRATION_DOCUMENT_TYPE_CODE = "F32",
    DPP_PREMISE_FORM_20_REGISTRATION_DOCUMENT_TYPE_CODE = "F34",

    DPP_PREMISE_FORM_18_RENEWAL_REGISTRATION_DOCUMENT_TYPE_CODE = "D74",
    DPP_PREMISE_FORM_19_RENEWAL_REGISTRATION_DOCUMENT_TYPE_CODE = "D75",
    DPP_PREMISE_FORM_20_RENEWAL_REGISTRATION_DOCUMENT_TYPE_CODE = "D76",

    DPP_FORM_18_PREMISES_REGISTRATION_AMENDMENT_CODE = "D77",
    DPP_FORM_19_PREMISES_REGISTRATION_AMENDMENT_CODE = "D78",
    DPP_FORM_20_PREMISES_REGISTRATION_AMENDMENT_CODE = "D79",

    //Catch Certificate
    CATCH_CERTIFICATE_DOCUMENT_TYPE_CODE = "CC1",
    CATCH_CERTIFICATE_DOCUMENT_CODE = "CC2",

    //Export Permit
    EXPORT_PERMIT_DOCUMENT_TYPE_CODE = "A22",
    EXPORT_PERMIT_DOCUMENT_CODE = "A23",
    //DPP PREMISE REGISTRATION
    DPP_FORM_18_PREMISES_REGISTRATION_REQUEST_CODE = "F30",
    DPP_FORM_18_PREMISES_REGISTRATION_CERTIFICATE_CODE = "F29",
    DPP_FORM_19_PREMISES_REGISTRATION_REQUEST_CODE = "F32",
    DPP_FORM_19_PREMISES_REGISTRATION_CERTIFICATE_CODE = "F31",
    DPP_FORM_20_PREMISES_REGISTRATION_REQUEST_CODE = "F34",
    DPP_FORM_20_PREMISES_REGISTRATION_CERTIFICATE_CODE = "F33",
    MNC_IMPORT_PERMIT_DOCUMENT_TYPE_CODE = "A20",
    MNC_IMPORT_PERMIT_COM_DOCUMENT_TYPE_CODE = "A21",
    
    // ePhyto Certificate 
    EPHYTO_EC_DOCUMENT_TYPE_CODE="EP1",

    // CESS WAIVER/EXEMPTION
    CESS_EXEMPTION_DOCUMENT_TYPE_CODE = "WA1",
    CESS_EXEMPTION_AGAINST_NTN_DOCUMENT_TYPE_CODE = "WA2",

    // Drug Sale License
    DRUG_SALES_LICENSE_REGISTRATION_TYPE_CODE = "D98",

    // Drug Sale License
    DRUG_MANUFACTURING_LICENSE_REGISTRATION_TYPE_CODE = "D92",
    //BANK GUARANTEE DOCUMENT
    BANK_GUARANTEE_DOCUMENT_TYPE_CODE = "GU2",
}

export enum DocumentClassificationCode {
    RELEASE_ORDER = "RO",
    IMPORT_PERMIT = "IMP",
    IMPORT_PERMIT_AMENDMENT = "IPA",
    EXPORT_CERTIFICATE = "EC",
    HEALTH_CERTIFICATE = "HC",
    SEED_ENLISTMENT = "PRD",
    PRODUCT = "PRR",
    PRODUCT_AMENDMENT = "PRA",
    BUSINESS = "BSS",
    PREMISE = "PRM",
    CATCH_CERTIFICATE = "CC",
    EXPORT_PERMIT = "EXP",
    ENTITY_REGISTRATION = "BSS",
    IMPORT_PERMIT_OPINION_REQUEST = "IPO",
    WAIVER = "WR",
    WAIVER_AGAINST_NTN = "WR2",
    DRUG_SALES_LICENSE = "DSL",
    DRUG_MANUFACTURING_LICENSE = "DML",
    GROUD_CHECK_REPORT = "GCR",
    GROUD_CHECK_REPORT_AMENDMENT = "GCR",
    LICENSE_REGISTRATION= "ALR",
    DRUG_REGISTRATION= "DDR",
    EPHYTO="EPH",
    TREATMENT_PROVIDER = 'TPM'
}
// ...
export enum Agency {
    DPP = 2,
    AQD = 3,
    FSCRD = 4,
    PSQCA = 5,
    PSI = 6,
    MFD = 10,
    MNC = 11,
    DRAP = 12,
    MMD = 13,
    IOCO = 14,
    SETN = 15
}
// ...
export enum DocumentGroupCode {
    IMPORT_PERMIT = "IP",
    RELEASE_ORDER = "RO",
    EXPORT_CERTIFICATE = "EC",
    SEED_ENLISTMENT = "PRD",
    PREMISE_REGISTARTION = "PRM",
    EXORT_PERMIT = "EXP",
    IMPORT_PERMIT_OPINION_REQUEST = "IPO",
    PRODUCT_REGISTARTION = "PRR",
    PRODUCT_REGISTARTION_AMENDMENT = "PRA",
    CESS_WAIVER = "CW",
    CESS_WAIVER_AGAINST_NTN = "CW2",
    PREMISE_REGISTARTION_AMENDMENT = "PMA",
    PREMISE_REGISTARTION_RENEWAL = "PMR",
    DRUG_SALES_LICENSE= "DSL",
    DRUG_MANUFACTURING_LICENSE= "DML"
}